requirements.txt
requests
pytest
python-dotenv
psycopg2
jsonschema
pytest-pyreport
